+++ src/grocker/__init__.py
@@ -7,4 +7,4 @@ import pkg_resources
 __version__ = pkg_resources.get_distribution('grocker').version
 __copyright__ = '2015, Polyconseil'
 
-DOCKER_API_VERSION = '1.21'
+DOCKER_API_VERSION = '1.23'